+++ Ragdoll/src/Ragdoll/GeometryBuilder.h
@@ -8,11 +8,11 @@ class GeometryBuilder {
 public:
 	void Init(nvrhi::DeviceHandle nvrhiDevice);
 	
-	int32_t BuildCube(float size);
-	int32_t BuildSphere(float diameter, uint32_t tessellation);
-	int32_t BuildCylinder(float height, float diameter, size_t tessellation);
-	int32_t BuildCone(float diameter, float height, size_t tessellation);
-	int32_t BuildIcosahedron(float size);
+	size_t BuildCube(float size);
+	size_t BuildSphere(float diameter, uint32_t tessellation);
+	size_t BuildCylinder(float height, float diameter, uint32_t tessellation);
+	size_t BuildCone(float diameter, float height, uint32_t tessellation);
+	size_t BuildIcosahedron(float size);
 private:
 	nvrhi::DeviceHandle Device;
 	std::vector<Vertex> Vertices;
